requirements.txt
pyirsdk==1.3.5
PyYAML==6.0.1
